move generation of server user properties to application
they used to be in a .env file. This is first of all a security
issue since an attacker can view password and private key when having
access to the filesystem and secondly a hassle for the user to setup
the application since a rake task must be executed before initially
starting the application.